Remove SQLite stuff from TrackCollection & Co
