+++ src/library/basetrackcache.cpp
@@ -32,8 +32,8 @@ BaseTrackCache::BaseTrackCache(TrackCollection* pTrackCollection,
           m_bIndexBuilt(false),
           m_bIsCaching(isCaching),
           m_trackDAO(pTrackCollection->getTrackDAO()),
-          m_database(pTrackCollection->getDatabase()),
-          m_pQueryParser(new SearchQueryParser(pTrackCollection->getDatabase())) {
+          m_database(pTrackCollection->database()),
+          m_pQueryParser(new SearchQueryParser(pTrackCollection->database())) {
     m_searchColumns << "artist"
                     << "album"
                     << "album_artist"
