+++ src/library/librarytablemodel.cpp
@@ -32,7 +32,7 @@ void LibraryTableModel::setTableModel(int id) {
 
     const QString tableName = "library_view";
 
-    QSqlQuery query(m_pTrackCollection->getDatabase());
+    QSqlQuery query(m_pTrackCollection->database());
     QString queryString = "CREATE TEMPORARY VIEW IF NOT EXISTS " + tableName + " AS "
             "SELECT " + columns.join(", ") +
             " FROM library INNER JOIN track_locations "
