+++ src/library/rhythmbox/rhythmboxfeature.cpp
@@ -57,7 +57,7 @@ RhythmboxFeature::RhythmboxFeature(QObject* parent, TrackCollection* pTrackColle
     m_isActivated =  false;
     m_title = tr("Rhythmbox");
 
-    m_database = QSqlDatabase::cloneDatabase(pTrackCollection->getDatabase(),
+    m_database = QSqlDatabase::cloneDatabase(pTrackCollection->database(),
                                              "RHYTHMBOX_SCANNER");
 
     //Open the database connection in this thread.
